+++ src/main/java/com/vluepixel/vetmanager/api/product/core/application/port/in/FindProductPort.java
@@ -22,5 +22,5 @@ public interface FindProductPort {
      * @param id the id.
      * @return product found
      */
-    ProductDto findById(Integer id);
+    ProductDto findById(Long id);
 }
